+++ lib/main.ts
@@ -216,6 +216,12 @@ export type {
   StackGap,
 } from './components/Stack/Stack';
 
+export { AutoGrid } from './components/AutoGrid/AutoGrid';
+export type {
+  AutoGridProps,
+  AutoGridGap,
+} from './components/AutoGrid/AutoGrid';
+
 export { Layout } from './components/Layout/Layout';
 export type { LayoutProps } from './components/Layout/Layout';